fix type error

+++ backend/koj-backend-server/src/main/kotlin/com/kairlec/koj/backend/controller/admin/ProblemManagerController.kt
@@ -7,6 +7,7 @@ import com.kairlec.koj.dao.tables.records.ProblemConfigRecord
 import com.kairlec.koj.dao.tables.records.ProblemRunRecord
 import kotlinx.coroutines.flow.Flow
 import org.springframework.web.bind.annotation.*
+import java.math.BigInteger
 
 @RestController
 @RequestMapping("/admin")
@@ -93,9 +94,9 @@ class ProblemManagerController(
         val languageId: String,
         val time: Int,
         val memory: Int,
-        val maxOutputSize: Long?,
-        val maxStack: Long?,
-        val maxProcessNumber: Short?,
+        val maxOutputSize: BigInteger?,
+        val maxStack: BigInteger?,
+        val maxProcessNumber: Int?,
         val args: List<String>?,
         val env: List<String>?
     )
